+++ src/Agent/NewRelic/Agent/Core/Configuration/DefaultConfiguration.cs
@@ -378,6 +378,21 @@ namespace NewRelic.Agent.Core.Configuration
             }
         }
 
+        private bool? _ignoreServerSideConfiguration;
+
+        public bool IgnoreServerSideConfiguration
+        {
+            get
+            {
+                if (!_ignoreServerSideConfiguration.HasValue)
+                {
+                    _ignoreServerSideConfiguration = EnvironmentOverrides(false, "NEW_RELIC_IGNORE_SERVER_SIDE_CONFIG");
+                }
+
+                return _ignoreServerSideConfiguration.Value;
+            }
+        }
+
         public bool AllowAllRequestHeaders => HighSecurityModeOverrides(false, _localConfiguration.allowAllHeaders.enabled);
 
         #region Attributes
@@ -1655,7 +1670,7 @@ namespace NewRelic.Agent.Core.Configuration
 
         public virtual IEnumerable<long> TrustedAccountIds { get { return _serverConfiguration.TrustedIds ?? new List<long>(); } }
 
-        public bool UsingServerSideConfig { get { return _serverConfiguration.UsingServerSideConfig; } }
+        public bool ServerSideConfigurationEnabled { get { return _serverConfiguration.ServerSideConfigurationEnabled; } }
 
         #region Metric naming