+++ src/Agent/NewRelic/Agent/Core/Configuration/ServerConfiguration.cs
@@ -79,7 +79,7 @@ namespace NewRelic.Agent.Core.Configuration
 
         // Server Side Config
 
-        public bool UsingServerSideConfig { get; private set; }
+        public bool ServerSideConfigurationEnabled { get; private set; }
 
         [JsonProperty("agent_config")]
         public AgentConfig RpmConfig { get; set; }
@@ -301,12 +301,17 @@ namespace NewRelic.Agent.Core.Configuration
             public IEnumerable<string> Terms { get; set; }
         }
 
-        public static ServerConfiguration FromJson(string json)
+        public static ServerConfiguration FromJson(string json, bool ignoreServerConfiguration = false)
         {
             var serverConfiguration = JsonConvert.DeserializeObject<ServerConfiguration>(json);
             Debug.Assert(serverConfiguration != null);
 
-            serverConfiguration.UsingServerSideConfig = JsonContainsNonNullProperty(json, "agent_config");
+            if (ignoreServerConfiguration)
+            {
+                serverConfiguration.RpmConfig = new AgentConfig();
+            }
+
+            serverConfiguration.ServerSideConfigurationEnabled = JsonContainsNonNullProperty(json, "agent_config");
 
             return serverConfiguration;
         }
@@ -320,10 +325,10 @@ namespace NewRelic.Agent.Core.Configuration
                 && dictionary[propertyName] != null;
         }
 
-        public static ServerConfiguration FromDeserializedReturnValue(object deserializedJson)
+        public static ServerConfiguration FromDeserializedReturnValue(object deserializedJson, bool ignoreServerConfiguration = false)
         {
             var json = JsonConvert.SerializeObject(deserializedJson);
-            return FromJson(json);
+            return FromJson(json, ignoreServerConfiguration);
         }
 
         [OnError]