+++ src/Agent/NewRelic/Agent/Core/DataTransport/ConnectionHandler.cs
@@ -233,19 +233,35 @@ namespace NewRelic.Agent.Core.DataTransport
 
             Log.InfoFormat("Agent {0} connected to {1}:{2}", GetIdentifier(), _connectionInfo.Host, _connectionInfo.Port);
 
-            var serverConfiguration = ServerConfiguration.FromDeserializedReturnValue(responseMap);
+            var serverConfiguration = ServerConfiguration.FromDeserializedReturnValue(responseMap, _configuration.IgnoreServerSideConfiguration);
             LogConfigurationMessages(serverConfiguration);
 
             return serverConfiguration;
         }
 
-        private static void LogConfigurationMessages(ServerConfiguration serverConfiguration)
+        private void LogConfigurationMessages(ServerConfiguration serverConfiguration)
         {
             if (serverConfiguration.HighSecurityEnabled == true)
+            {
                 Log.Info("The agent is in high security mode.  No request parameters will be collected and sql obfuscation is enabled.");
+            }
+
+            if (serverConfiguration.ServerSideConfigurationEnabled)
+            {
+                if (_configuration.IgnoreServerSideConfiguration)
+                {
+                    Log.Info("Server-Side Configuration is enabled, but the agent is configured to ignore it.");
+                }
+                else
+                {
+                    Log.Info("Server-Side Configuration is enabled.");
+                }
+            }
 
             if (serverConfiguration.Messages == null)
+            {
                 return;
+            }
 
             foreach (var message in serverConfiguration.Messages)
             {
@@ -328,7 +344,8 @@ namespace NewRelic.Agent.Core.DataTransport
                 EncodingKey = _configuration.EncodingKey,
                 TrustedAccountIds = _configuration.TrustedAccountIds.ToList(),
                 MaxStackTraceLines = _configuration.StackTraceMaximumFrames,
-                UsingServerSideConfig = _configuration.UsingServerSideConfig,
+                ServerSideConfigurationEnabled = _configuration.ServerSideConfigurationEnabled,
+                IgnoreServerSideConfiguration = _configuration.IgnoreServerSideConfiguration,
                 ThreadProfilerEnabled = _configuration.ThreadProfilingEnabled,
                 CrossApplicationTracerEnabled = _configuration.CrossApplicationTracingEnabled,
                 DistributedTracingEnabled = _configuration.DistributedTracingEnabled,