+++ tests/Agent/IntegrationTests/SharedApplications/Common/MultiFunctionApplicationHelpers/NetStandardLibraries/NServiceBus/Models/Event.cs
@@ -1,0 +1,14 @@
+// Copyright 2020 New Relic, Inc. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+using System;
+using NServiceBus;
+
+#if !NET462
+
+namespace NsbTests
+{
+    public class Event : BaseModel, IEvent { }
+}
+
+#endif