+++ tests/Agent/IntegrationTests/SharedApplications/Common/MultiFunctionApplicationHelpers/NetStandardLibraries/NServiceBus5/NServiceBusService.cs
@@ -1,0 +1,77 @@
+﻿// Copyright 2020 New Relic, Inc. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+
+#if NET462
+
+using NewRelic.Api.Agent;
+using NewRelic.Agent.IntegrationTests.Shared.ReflectionHelpers;
+using NServiceBus;
+using System;
+using System.Collections.Generic;
+using System.Runtime.CompilerServices;
+
+namespace MultiFunctionApplicationHelpers.NetStandardLibraries.NServiceBus5
+{
+    [Library]
+    public class NServiceBusService
+    {
+        private const string DestinationReceiverHost = "NServiceBusReceiverHost";
+        private static IBus _bus;
+        private static Random _random;
+
+        [LibraryMethod]
+        public void Start()
+        {
+            Logger.Info($"Starting NServiceBusService");
+            var busConfig = new BusConfiguration();
+            busConfig.UsePersistence<InMemoryPersistence>();
+            var typeToScan = new List<Type>
+            {
+                typeof(SampleNServiceBusMessage),
+                typeof(SampleNServiceBusMessage2)
+            };
+            busConfig.TypesToScan(typeToScan);
+            busConfig.CustomConfigurationSource(new ConfigurationSource());
+            _bus = NServiceBus.Bus.Create(busConfig);
+            _random = new Random();
+            Logger.Info($"NServiceBusService Started");
+        }
+
+        [LibraryMethod]
+        public void Stop()
+        {
+            Logger.Info($"Stopping NServiceBusService");
+            _bus.Dispose();
+            Logger.Info($"NServiceBusService Stopped");
+        }
+
+        [LibraryMethod]
+        [Transaction]
+        [MethodImpl(MethodImplOptions.NoOptimization | MethodImplOptions.NoInlining)]
+        public void Send()
+        {
+            var message = new SampleNServiceBusMessage(_random.Next(), "Foo bar");
+            _bus.Send(DestinationReceiverHost, message);
+            Logger.Info($"Message with ID={message.Id} sent via NServiceBus" );
+        }
+
+        [LibraryMethod]
+        public void SendValid()
+        {
+            var message = new SampleNServiceBusMessage2(_random.Next(), "Valid");
+            _bus.Send(DestinationReceiverHost, message);
+            Logger.Info($"Message with ID={message.Id} sent via NServiceBus");
+        }
+
+        [LibraryMethod]
+        public void SendInvalid()
+        {
+            var message = new SampleNServiceBusMessage2(_random.Next(), "Invalid", false);
+            _bus.Send(DestinationReceiverHost, message);
+            Logger.Info($"Message with ID={message.Id} sent via NServiceBus");
+        }
+    }
+}
+
+#endif